Struct binary_tree::count::CountNode

pub struct CountNode<T> {
    // some fields omitted
}

Node of a CountTree.

The only way of getting your hands on a CountNode is through CountTree::root() method which returns a shared reference to its root. Thus NodeMut methods are not accessible to users.