Module binary_tree::count

Counting tree implementation.

When should you use CountTree?

You want to maintain a possibly large unsorted list.

You want to access, modify, insert, and delete elements at arbitrary position with O(log(n)) time complexity.

You can tolerate O(n log(n)) time-complexity for (not implemented yet):

splitting at arbitrary position

truncating the length (complexity unclear)

appending another list (complexity unclear)

You have less than 4.29 billion (u32::MAX) elements!

Structs

CountNode
Node of a CountTree.

CountTree
Counting tree.

IntoIter

Iter

Type Definitions

NodePtr